se escribe mapeo en archivos .csv

+++ coordinador.h
@@ -9,7 +9,7 @@ typedef struct
 typedef struct
 {
   char grupo_vehiculo[20];
-  float tasacion;
-  float valor_pagado;
+  int tasacion;
+  int valor_pagado;
   int puertas;
 } Vehiculo;
